Update file_server.c
Using this code for a fileserver on a sd-card with subdirectories failes because "opendir" behaves differently for the mountpoint and deeper layoing subdirectories. For subdirectories, there must not be a "/" at the end - differently than to the mountpoint:
mountpoint: opendir ("/sdcard/")
subdirectory: opendir("/sdcard/subdir")

Therefore I have added a if-clause, that removes the last "/", if it is not the mountpoint.